testapi/requirements.txt
asgiref==3.7.2
automium==0.2.6
automium-web==0.1.1
beautifulsoup4==4.12.2
blinker==1.6.2
cheroot==10.0.0
click==8.1.7
Django==4.2.4
django-appconf==1.0.5
django-bootstrap-form==3.4
django-bootstrap3==23.4
django-bootstrap4==23.2
django-forms-bootstrap==3.1.0
django-user-accounts==3.2.0
Flask==2.3.2
itsdangerous==2.1.2
jaraco.functools==3.8.1
Jinja2==3.1.2
legacy==0.1.6
MarkupSafe==2.1.3
more-itertools==10.1.0
Pinax==0.9a2
pinax-theme-bootstrap==8.0.1
psycopg2-binary==2.9.7
pytz==2023.3
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.7.1
web.py==0.62
Werkzeug==2.3.7
